refactor(runtime): consolidate sweep/security/submission_plan under runtime/ (#164)
Phase 8f of the src/srunx/ cleanup (#156). Physical-co-location move so
every module the architecture guard already maps to the ``runtime``
layer now lives under ``src/srunx/runtime/``.

Moves:
  - src/srunx/sweep/                 → src/srunx/runtime/sweep/
  - src/srunx/security/              → src/srunx/runtime/security/
  - src/srunx/cli/submission_plan.py → src/srunx/runtime/submission_plan.py

Shims preserve legacy import paths:
  - ``srunx.sweep.*`` aliases every submodule via sys.modules.
  - ``srunx.security.*`` aliases every submodule via sys.modules.
  - ``srunx.cli.submission_plan`` aliases as a whole module.

Consumer updates (switching to the canonical path):
  - ``src/srunx/slurm/ssh.py``
  - ``src/srunx/web/routers/jobs.py``
  - ``src/srunx/web/routers/workflows.py``
  - ``src/srunx/cli/workflow.py``, ``src/srunx/cli/main.py``

Architecture guard wins:
  - ``(slurm, cli)`` KNOWN_VIOLATION removed — SSH adapter now reaches
    into ``runtime.submission_plan`` (runtime layer, allowed).
  - ``(sweep, callbacks)`` / ``(sweep, db)`` / ``(sweep, notifications)``
    folded into ``(runtime, callbacks)`` / ``(runtime, db)`` / new
    ``(runtime, notifications)`` entry — same edges, new source package.

Behavior unchanged. 1978 / 1981 tests pass; the 3 failing tests are
pre-existing env-specific issues.
