+++ src/srunx/runtime/security/__init__.py
@@ -1,0 +1,17 @@
+"""Security-related helpers shared across transport boundaries (Web, MCP)."""
+
+from srunx.runtime.security.mount_paths import (
+    ShellJobScriptViolation,
+    find_shell_script_violation,
+)
+from srunx.runtime.security.python_args import (
+    PythonPrefixViolation,
+    find_python_prefix,
+)
+
+__all__ = [
+    "PythonPrefixViolation",
+    "ShellJobScriptViolation",
+    "find_python_prefix",
+    "find_shell_script_violation",
+]
